added CoordMath Class
